Feat: VAM API call
Added an API call to the Victoria and albert museum to get their collection.
Added a call to this function to a useEffect on app.tsx so that it is only called when the app launches.
Investigated using the met museum of art API aswell (will likely be by department since the number of item ids returned from calling objects was 5 figures).
Added ArtObject interface to ensure that the data is properly structured and a Convert_to_ArtObject function to take the entries from each API and convert them to ArtObjects.
Added some sample API responses, to avoid calling the real thing dozens of times and getting blocked.
Updated API sugestions.txt