Feat: Call API and convert to ArtObject
Added a function to call the VAM API and return a promise with a ready to use list of art objects
Decided on the fields to be used in an ArtObject(the standardised object we will be using to display data from the different APIs).
Added a conversion for data from the VAM API to the ArtObject format.
Added some testing code to App.tsx to ensure that the calls work and can be displayed (this will likely become a context provider if I choose to make the data available app wide)