Feat: search results page and objects
Added a search results component that takes an art object and display a summary of it, along with the "add to exhibition" button.
Added a search results page to display said result objects.
Added a placeholder title to Convert_to_ArtObject.tsx to use if the title is blank.
Minor improvement to homepage.tsx and notfound.tsx.