Feat: search prototype
Added Run_search.tsx, with both a normal function and a functional component as methods of running the search, both need some refinement(rendering the output on a new page (tried router, but that's alot of data for a paramater. or calling the FC properly via the search bar)
Put a demo of the finished search on app.tsx for now (will remove when the function is on it's own page).
Added startsearch function to the searchbar's search button, which attempts to render a searchResultsPage using the data from search_apis
Added an indicator for no results to the search results page.
Changed the serach strings on fetch_met_API and fetch vam_API to allow the search term to be set via a paramater.