Feat: context and searchbar completion
Added an app context object that currently stores the last search details (will likely include information on exhibition contents aswell).
Updated the searchbar to set it's contents as the new search criteria and move to the search page (or tell you to enter something).
Updated search results page to grab the search criteria and prepare to run the search (looks like some API value names will require fixing), it should now run the search and, when fixed, map the results onto the page.
Changed the router and todo.txt to reflect the new plan
Updated IsNotBlank to cover more usecases