Feat: fully functional search page
Modified the search results page to correctly call search_APIs and to display a "searching" message while waiting, and a result count/ no results message as appropriate.
Modified Run_search's search_APIs function to return a promise with it's art object array, rather than just the array, so that it can be called aysnc and properly used to set state.