Style: Inital responsive design
Made modifications to several elements to make them more responsive.
Modified elements include: the header, the search bar(might need an intermediate step), exhibition container and exhibition object
worked on adding an added/removed animation to the search result (still needs to be responsive.
Added a gap between the end of content and the footer