Feat: Search result organisation
Added some code to move the search results with unknown artists and titles to the bottom of the results list.
Changed placeHolderArtOBJ to include some more descriptive values.
Fixed a small css mistake in the search result image.
Replaced the typed out placeholder in data in Convert_to_ArtObject with the placeholder object.